backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
google-genai==1.70.0
anthropic==0.28.0
python-docx==1.1.2
reportlab==4.4.0
gunicorn==23.0.0
